+++ src/test/java/com/hunchee/twist/util/EntityUtilTest.java
@@ -25,12 +25,15 @@ package com.hunchee.twist.util;
 import com.google.appengine.api.datastore.DatastoreService;
 import com.google.appengine.api.datastore.DatastoreServiceFactory;
 import com.google.appengine.api.datastore.Entity;
-import com.google.appengine.repackaged.com.google.common.collect.Lists;
 
 import com.hunchee.twist.LocalDatastoreTestBase;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
+
+import java.util.ArrayList;
+import java.util.Arrays;
+
 import static org.junit.Assert.*;
 
 /**
@@ -62,7 +65,7 @@ public class EntityUtilTest extends LocalDatastoreTestBase {
     
     Entity e2 = new Entity("baz", "Bob", e1.getKey());
     Entity e3 = new Entity("foo");
-    ds.put(Lists.newArrayList(e2, e3));
+    ds.put(new ArrayList<Entity>(Arrays.asList(e2, e3)));
     
     assertTrue(EntityUtil.inSameEntityGroup(ancestor, e1, e2));
     assertFalse(EntityUtil.inSameEntityGroup(ancestor, e1, e2, e3));